requirements.txt
fastapi
fastapi[standard]
PyJWT
pydantic
mysqlclient
sqlalchemy